remove RefWatchingController, create RefWatcherDelegate
