+++ app/src/main/java/net/yslibrary/monotweety/base/BaseController.kt
@@ -5,6 +5,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import com.bluelinelabs.conductor.ControllerChangeHandler
+import com.bluelinelabs.conductor.ControllerChangeType
+import com.bluelinelabs.conductor.rxlifecycle.RxController
+import com.squareup.leakcanary.RefWatcher
 import net.yslibrary.monotweety.analytics.Analytics
 import net.yslibrary.monotweety.base.di.Names
 import net.yslibrary.rxeventbus.EventBus
@@ -18,7 +22,8 @@ import kotlin.properties.Delegates
 /**
  * Created by yshrsmz on 2016/09/24.
  */
-abstract class BaseController() : RefWatchingController() {
+abstract class BaseController() : RxController(),
+                                  RefWatcherDelegate by RefWatcherDelegateImpl() {
 
   @set:Inject
   @setparam:Named(Names.FOR_ACTIVITY)
@@ -26,6 +31,9 @@ abstract class BaseController() : RefWatchingController() {
 
   @set:Inject
   var analytics by Delegates.notNull<Analytics>()
+
+  @set:[Inject]
+  var refWatcher: RefWatcher? = null
 
   private var created: Boolean = false
 
@@ -38,6 +46,7 @@ abstract class BaseController() : RefWatchingController() {
     if (!created) {
       created = true
       onCreate()
+      takeRefWatcher(refWatcher)
     }
 
     return inflateView(inflater, container)
@@ -48,6 +57,11 @@ abstract class BaseController() : RefWatchingController() {
   }
 
   abstract fun inflateView(inflater: LayoutInflater, container: ViewGroup): View
+
+  override fun onChangeEnded(changeHandler: ControllerChangeHandler, changeType: ControllerChangeType) {
+    super.onChangeEnded(changeHandler, changeType)
+    onChangeEnded(isDestroyed, changeType)
+  }
 
   fun <T> Observable<T>.bindToLifecycle(): Observable<T> = this.compose(this@BaseController.bindToLifecycle<T>())
